Fix spurious installation warning
Fix spurious installation warning claiming that MATLAB's
"Symbolic Math Toolbox" isn't installed even if it is.

The bug was caused by the fact that when searching for the toolbox
name in the output of the `ver` command to assess whether the
toolbox is installed, a trailing blank space was included in the
string to match (i.e. the toolbox name). This commit removes the
trailing blank space.